Ensure context ID can be set in bootstrap.yml
The problem was the ordering of ApplicationContextInitializers:
the bootstrap process sets the parent application context in an
initializer (formerly with no @Order) which needs to happen
before any other initializers that depend on the Environment.
Fix is to add an @Order (e.g. 0) so it comes before the
ContextIdApplicationContextInitializer.

Fixes gh-37