Add listener for HearbeatEvent in discovery client config boostrap
The HeartbeatEvent can tell us when the config server might have moved,
so we listen for it and refresh the configuration if there is a
change in the service catalog. On startup if there is a retry
configured it will catch the change. After that the app has to be
/refresh'ed to propagate the change to the environment.

Fixes gh-400