Issue #177: Alternative implementation of Source Locator: Authorization
This commit adds the Cloud Config Property "authorization", as the HTTP
Authorization header to connect to the client. Users won't be able to
set both the "password" and "authorization" at the same time.

When set, the client will use the value to create a generic
Authorization header with the value provided.

Discussions about this at
https://github.com/spring-cloud/spring-cloud-config/issues/177#issuecomment-239076005
*	modified:   spring-cloud-config-client/src/main/java/org/springframework/cloud/config/client/ConfigClientProperties.java
- Adding the property "Authorization"
- Adding setter/getter
- Updating the toString() to contain the authorization

*	modified:   spring-cloud-config-client/src/main/java/org/springframework/cloud/config/client/ConfigServicePropertySourceLocator.java
- Validating the set of either the password or authorization
- Creating the new class GenericAuthorization.