Alternative implementation of Source Locator: Authorization
This commit adds the Cloud Config Property "authorization", as the HTTP
Authorization header to connect to the client. Users won't be able to
set both the "password" and "authorization" at the same time.

When set, the client will use the value to create a generic
Authorization header with the value provided.

Fixes gh-474