Slightly less bizarre implementation of convertToProperties()
The EnvironmentController has to convert an Environment to
Properties in a few places. We have had issues in the past
with inconsistent behaviour for arrays (e.g. a.b[0], a.b[1])
defined and overridden in more than one property source. The
fix for those issues created other problems because too many
of the other properties in the property source that overrides
an array were carried over into the final result.

Fixes gh-441