Additional tests for placeholder replacement
It turns out that the mechanism we use to convert properties
to YAML (and JSON) means that placeholders are always replaced
if there are defaults (e.g. ${foo:bar} always resolves to "bar"
if "foo" is undefined).

This seems soprt of reasonable, if a little surprising. We could
change the way YAML and JSON are generated, but that would be
a big deal, and probably not worth the effort to fix a bit of
slightly surprising behaviour.