Fixing a typo in the documentation
There's a typo in the *Vault Server* section:
When using Vault as a backend you can share configuration with all applications by placing configuration in `html5/application`. For example, if you run this Vault command

When it should be:
When using Vault as a backend you can share configuration with all applications by placing configuration in `secret/application`. For example, if you run this Vault command