Derive base dir of new pattern matching repositories from the parent
Before this change when a repository is created to support a pattern match
in the URL, the basedir for cloning the scm repository is a new temp dir.
So if user configures the basedir to be something other than temp then he
loses the benefit once patterns are used in the URL matchers.

The basedir still has to be unique, so we now use the same parent directory
as the parent repo, and add the temp file name as a directory name. I.e.
if basedir=/foo/bar the pattern matched basedir will be something like
/foo/config-repo-23874691. You need the directory to have a parent that
is writable by the config server JVM.

Fixes gh-451