Check for index.lock before attempting to use JGit
A stale lock from a crashed JVM can prang the JGit client by
preventing it from ever getting access to the working directory.
This change attempts to clean the slate by removing the lock -
because of the synchronization already present in the repository
the only way this can go wrong is if multiple server instances
are trying to share the working directory (which would be an
eror for other reasons).