Pull config properties out of environment repositories
In order to support multiple auto-config approaches to EnvironmentRepository configuration, we need to get the ConfigurationProperties annotation off of the EnvironmentRepository implementations. As a result, the getters and setters of these properties won't need to be on the repo implementations either. Marking them as Deprecated for now, to maintain backwards compatibility.