Add list-based composite auto-configuration (#912)
* Pull config properties out of environment repositories

In order to support multiple auto-config approaches to EnvironmentRepository configuration, we need to get the ConfigurationProperties annotation off of the EnvironmentRepository implementations. As a result, the getters and setters of these properties won't need to be on the repo implementations either. Marking them as Deprecated for now, to maintain backwards compatibility.

* Add list-based composite auto-configuration

Add support for using a list of environment repository configuration blocks, under spring.cloud.config.server.composite, to create a composite environment repository.