#974 remove one or two escapes for supporting JSON (#982)
Hi.  This is an attempt to fix #974 (cc @ryanjbaxter )

As mentioned previously in the ticket, the original code tries to mask out escaped variables (i.e. `\${hello}`) before performing replacement, and then finally removing the escapes before returning.  

This works fine for .yaml or .properties, but not for JSON.

The jackson serialization that happens before the property resolution ends up adding an extra `\` (so that the JSON is well-formed) [EnvironmentController.labelledJsonPropertes](https://github.com/spring-cloud/spring-cloud-config/blob/master/spring-cloud-config-server/src/main/java/org/springframework/cloud/config/server/environment/EnvironmentController.java#L169).

I've updated the `resolverPlaceholders` to accept one or two escapes (`\` or `\\`).

I realize this will now treat `\\` as an escape in yml or properties, which may be unintended, but seems unlikely.

An alternative might be to add a flag to this method to indicate whether or not to allow double escape, and updating the `labelledJson` method to pass the flag?

I'm open to any feedback on this approach or another.

Thanks!