Add abstraction for retrieving config token.
The VaultEnvironmentRepository requires a token to be passed in a X-Config-Token header in an HttpServletRequest. This doesn't work well when config server is embedded in an application, as an HttpServletRequest is not available in that case.

These changes allows the Vault token to be provided in the HttpServletRequest header (the default behavior) or as a configuration property.