Add index regen if config server has force-pull
jGit may fail with 'JGitInternalException: Short read of block.'
Due to a truncated git index, which may happen due to external
reasons.
The index can be regenerated easily by deleting it and resetting
git working directory. For a pull force, resetting to HEAD is
more convenient since the info that may have been lost in the
index is the diff in a dirty copy that we are discarding in
any way.

This commits adds automates those steps when git-status fails
and the force-pull flag is set.

Please see:
- https://github.com/spinnaker/spinnaker/issues/5153
- https://bugs.eclipse.org/bugs/show_bug.cgi?id=531807
- https://www.devguerrilla.com/notes/2015/07/fuse-for-the-idle-fellow-short-read-of-block-errors-when-starting-a-container/