Updates ConfigServerHealthIndicator.java for ConfigData
ConfigClientAutoConfiguration now uses @ConditionalOnEnabledHealthIndicator("config") to enable the bean.

The configClient property source is now always added (even if empty) on successful call to config server.

The ConfigServerHealthIndicator now uses the existence of the configClient property source as the indicator that config server is UP. No new calls to config server are made by the health indicator.

Fixes gh-1696