Adds ConfigData support for ConfigServerInstanceProvider.Function
DiscoveryClient implementation can now register a ConfigServerInstanceProvider.Function using a Bootstrapper registered in spring.factories.

If spring.cloud.config.discovery.enabled is true, config client will use the above function to locate the initial instance of config server. A ConfigServerInstanceMonitor will be promoted to the application context so subsequent updates to the instanecs of config server will be reported.