Fix AWS S3 ResourceController support. (#2219)
There are several problems here:

* Fixes and improves documentation
* Fixes a bug where we were using the application as the bucket folder name when it should have been the label
* Uses the default label property if set and the query parameter useDefaultLabel is supplied
* Fixes the resource location check so it takes into account the encoded / in the path name of an S3 object

There is one use case that requires a fix in Spring Cloud AWS, and that is when there is no label supplied and there is no default label set.  There appears to be a. bug in the way the createRelative method creates the S3 object name in this case.  I have submitted a PR for this https://github.com/awspring/spring-cloud-aws/issues/648

Fixes #2216

Co-authored-by: Ryan Baxter <524254+ryanjbaxter@users.noreply.github.com>