Do not return profile-specific property sources
When a property source uses spring.profiles.import the config server would return property sources activated by this configuration property even if the config client did not request profile information.  This is due to using Spring Boots ConfigDataEnvironmentPostProcessor to load configuration from the file system (which is not only used for native but also for git and svn).  This change makes it so the config server will not activate profiles if the config client does not request any profile data.